+++ app/controllers/wishes_controller.rb
@@ -11,10 +11,13 @@ class WishesController < ApplicationController
 
   def destroy
     wish = Wish.find(params[:id])
-    activity = PublicActivity::Activity.where(trackable_type: "Wish").find_by(trackable_id: wish.id)
-    activity.destroy
+    # certains wish ont été crée avant qu'il n'y ait eu de publicActivity s'appliquant à eux
+    if PublicActivity::Activity.where(trackable_type: "Wish").find_by(trackable_id: wish.id).length > 0
+      activity = PublicActivity::Activity.where(trackable_type: "Wish").find_by(trackable_id: wish.id)
+      activity.destroy
+    end
     wish.destroy
-    redirect_to :back, notice: 'Le restaurant a bien été retirée de la liste de vos envies'
+    redirect_to :back, notice: 'Le restaurant a bien été retiré de votre wishlist'
   end
 
   private
